+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 ### Added
+- Added `mlockall` and `munlockall`
+  ([#876](https://github.com/nix-rust/nix/pull/876))
 - Added `SO_MARK` on Linux.
 - ([#873](https://github.com/nix-rust/nix/pull/873))
 - Added `getsid` in `::nix::unistd`